requirements.txt
earthaccess 
xarray 
netCDF4 
pandas 
geopandas 
psycopg2-binary 
python-dotenv 
